Also move utility functions to seperate file
